Title: 이차원 배열의 오름차순 정리, Time: 76ms, Memory: 8MB, Status: Wrong Answer - CodeTree

+++ 241023/이차원 배열의 오름차순 정리/ascending-order-of-two-dimensional-array.java
@@ -11,8 +11,8 @@ public class Main {
         n = Integer.parseInt(br.readLine());
         k = Integer.parseInt(br.readLine());
 
-        int left = 1;
-        int right = n * n;
+        long left = 1;
+        long right = n * n;
         long ans = n * n;
 
         while (left <= right) {
